chanages15sept

+++ lib/models/productsModel.dart
@@ -1,7 +1,7 @@
 class ProductModel {
   String productImage;
   String productName;
-  double productPrices;
+  int productPrices;
   ProductModel(
       {this.productImage = '', this.productName = '', this.productPrices = 0});
 }
